Convert GlyphRay to new dat format

+++ src/glyph_rays.rs
@@ -1,12 +1,12 @@
 pub struct GlyphRays{
-    pub width:i32,
-    pub height:i32,
-    pub l2r: Vec<i32>,
-    pub t2b: Vec<i32>,
-    pub r2l: Vec<i32>,
-    pub b2t: Vec<i32>,
-    pub m2l: Vec<i32>,
-    pub m2t: Vec<i32>,
-    pub m2r: Vec<i32>,
-    pub m2b: Vec<i32>
+    pub width:u16,
+    pub height:u8,
+    pub l2r: Vec<u16>,
+    pub t2b: Vec<u16>,
+    pub r2l: Vec<u16>,
+    pub b2t: Vec<u16>,
+    pub m2l: Vec<u16>,
+    pub m2t: Vec<u16>,
+    pub m2r: Vec<u16>,
+    pub m2b: Vec<u16>
 }
